+++ Source/DataBlock.swift
@@ -18,7 +18,7 @@ public struct DataBlock {
     public let icon: Icon?
     
     /// `DataPoint`s ordered by time, which describe the weather conditions at the requested location over time.
-    public let data: Array<DataPoint>
+    public let data: [DataPoint]
     
     /// Creates a new `DataBlock` from a JSON object.
     ///
@@ -26,18 +26,14 @@ public struct DataBlock {
     ///
     /// - returns: A new `DataBlock` filled with data from the given JSON object.
     public init(fromJSON json: NSDictionary) {
-        if let jsonSummary = json["summary"] as? String {
-            summary = jsonSummary
-        } else {
-            summary = nil
-        }
+        summary = json["summary"] as? String
         if let jsonIcon = json["icon"] as? String {
             icon = Icon(rawValue: jsonIcon)
         } else {
             icon = nil
         }
         
-        let jsonData = json["data"] as! Array<NSDictionary>
+        let jsonData = json["data"] as! [NSDictionary]
         var tempData = [DataPoint]()
         for jsonDataPoint in jsonData {
             tempData.append(DataPoint(fromJSON: jsonDataPoint))