+++ Source/Flag.swift
@@ -15,25 +15,25 @@ public struct Flag {
     public let darkSkyUnavailable: Bool?
     
     /// Contains the IDs for each radar station used in servicing the `Forecast` request.
-    public let darkSkyStations: Array<String>?
+    public let darkSkyStations: [String]?
     
     /// Contains the IDs for each `DataPoint` station used in servicing this request.
-    public let dataPointStations: Array<String>?
+    public let dataPointStations: [String]?
     
     /// Contains the IDs for each ISD station used in servicing this request.
-    public let isdStations: Array<String>?
+    public let isdStations: [String]?
     
     /// Contains the IDs for each LAMP station used in servicing this request.
-    public let lampStations: Array<String>?
+    public let lampStations: [String]?
     
     /// Contains the IDs for each METAR station used in servicing this request.
-    public let metarStations: Array<String>?
+    public let metarStations: [String]?
     
     /// The presence of this property indicates that data from api.met.no was used to facilitate this request (as per their license agreement).
     public let metnoLicense: Bool?
     
     /// Contains the IDs for each data station used in servicing this request.
-    public let sources: Array<String>
+    public let sources: [String]
     
     /// The presence of this property indicates which units were used for the data in this request. `US` units are default.
     public let units: Units
@@ -45,13 +45,13 @@ public struct Flag {
     /// - returns: A new `Flag` filled with data from the given JSON object.
     public init(fromJSON json: NSDictionary) {
         darkSkyUnavailable = json["darksky-unavailable"] as? Bool
-        darkSkyStations = json["darksky-stations"] as? Array<String>
-        dataPointStations = json["datapoint-stations"] as? Array<String>
-        isdStations = json["isd-stations"] as? Array<String>
-        lampStations = json["lamp-stations"] as? Array<String>
-        metarStations = json["metar-stations"] as? Array<String>
+        darkSkyStations = json["darksky-stations"] as? [String]
+        dataPointStations = json["datapoint-stations"] as? [String]
+        isdStations = json["isd-stations"] as? [String]
+        lampStations = json["lamp-stations"] as? [String]
+        metarStations = json["metar-stations"] as? [String]
         metnoLicense = json["metno-license"] as? Bool
-        sources = json["sources"] as! Array<String>
+        sources = json["sources"] as! [String]
         units = Units(rawValue: json["units"] as! String)!
     }
 }